Refactor code for conciseness: A file for generating weekly entries, another for results generation
